cleanup : custom commands
