+++ cmd/custom/deprecate/template.go.tmpl
@@ -78,7 +78,7 @@
 <main>
 
         <h1>Deprecated API(s) Dashboard </h1>
-        <p>Following the packages and its finding(s) obtained by checking the image and the bundle manifests distributed. This report aims to try to identify the packages distributions that can impact the users on 4.9.</p>
+        <p>Following the packages and its finding(s) obtained by checking the image and the bundle manifests distributed on it. This report aims to try to identify the packages distributions that can impact the users on 4.9.</p>
 
         <div class="container-fluid themed-container">
             <h5 class="display-12 fw-bold">Data from the image used</h5>
