+++ pkg/reports/custom/flags.go
@@ -1,7 +1,7 @@
 // Copyright 2021 The Audit Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
+// you may not use this File except in compliance with the License.
 // You may obtain a copy of the License at
 //
 //     http://www.apache.org/licenses/LICENSE-2.0
@@ -14,22 +14,10 @@
 
 package custom
 
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/operator-framework/audit/cmd/custom/deprecate"
-)
-
-func NewCmd() *cobra.Command {
-	indexCmd := &cobra.Command{
-		Use:   "dashboard",
-		Short: "generate specific custom reports based on the audit JSONs output",
-	}
-
-	indexCmd.AddCommand(
-		deprecate.NewCmd(),
-	)
-
-	return indexCmd
-
+// BindFlags define the Flags used to generate the bundle report
+type BindFlags struct {
+	File       string `json:"file"`
+	OutputPath string `json:"outputPath"`
 }
+
+var Flags = BindFlags{}
